requirements.txt
opencv-python~=4.10.0.84
matplotlib~=3.9.2
scikit-learn~=1.5.2
scikit-image~=0.24.0
torch~=2.4.1
python-dotenv~=1.0.1
joblib~=1.4.2
numpy~=2.1.2
seaborn~=0.13.2
PyYAML~=6.0.2
torchvision~=0.19.1